Some changes I did a while ago idk
